Adding even, odd partitioning teaser
Teaser: Given a variable length array of integers, partition them such that the even integers precede the odd integers in the array. Your
must operate on the array in-place, with a constant amount of extra space. The answer should scale linearly in time with respect to the size
of the array.